19 tests working

+++ src/main/java/Controller/SocialMediaController.java
@@ -1,7 +1,5 @@
 package Controller;
 
-import static org.mockito.ArgumentMatchers.contains;
-
 import Model.Account;
 import Model.Message;
 import Service.AccountService;
@@ -49,19 +47,20 @@ public class SocialMediaController {
     private void register(Context context) {
     
         Account account = context.bodyAsClass(Account.class);
-        int passLength = account.getPassword().length();
+        String password = account.getPassword();
         String uName = account.getUsername();
-        Account newAccount = aService.createAccount(account);
+        boolean exists = aService.accountExists(uName);
 
-        if(uName != null && passLength >= 4){
-            context.status(200);
+        if(uName != null && password.length() >= 4 && !uName.isBlank() && exists== false){
+            Account newAccount = aService.createAccount(uName,password);
+            if (newAccount != null){
+            context.status(200);}
             context.json(newAccount);
-        }{
+        }else {
             context.status(400);
         }
-        
-
     }
+
     private void login(Context context){
         Account account = context.bodyAsClass(Account.class);
         Account accountInfo = aService.login(account.getUsername(),account.getPassword());
@@ -75,29 +74,43 @@ public class SocialMediaController {
         
        //TODO: Finish LoginHandler
     }
+
     private void createMessage(Context context){
         Message message = context.bodyAsClass(Message.class);
         int messageLength = message.message_text.length();
 
-
         if(message != null && messageLength > 255){
-            Message newMessage = mService.insertMessage(message);
-            context.json(newMessage);
-            context.status(200);
-        }else{context.status(400);}
+
+            int postedByCount = mService.getPostedByCount(messageLength);
+            if(postedByCount > 0){
+                Message newMessage = mService.insertMessage(message);
+                context.json(newMessage);
+                context.status(200);
+
+            }}else{context.status(400);}
+            
+
 
     }
+    
     private void deleteMessage(Context context){
         
         int message_id = Integer.parseInt(context.pathParam("message_id"));
-        mService.deleteMessageById(message_id);
-
-        //TODO: Finish method
+        Message message = mService.getMessageById(message_id);
+        if(message!=null){
+            mService.deleteMessageById(message_id);
+            context.json(message);
+            context.status(200);}
+        else{
+            context.status(200);
+        }
 
     }
+
     private void getAllMessages(Context context){
         context.json(mService.getAllMessages());
     }
+
     private void getMessageByID(Context context){
         int message_id = Integer.parseInt(context.pathParam("message_id"));
         Message message = mService.getMessageById(message_id);
@@ -109,18 +122,24 @@ public class SocialMediaController {
         }
 
     }
+
+
     private void updateMessageByMessageID(Context context){
-        int message_id = Integer.parseInt(context.pathParam("message_id"));
-        int count = mService.getMessageCount(message_id);
-        String change = context.body();
-        if (change.length() < 255 && count>0){
-            Message updatedMessage = mService.updateMessage(message_id,change);
-            context.json(updatedMessage);
-            context.status(200);
-        }else{
-            context.status(400);
-        }
-    }
+         int message_id = Integer.parseInt(context.pathParam("message_id"));
+        Message existingMessage = mService.getMessageById(message_id); 
+        if (existingMessage != null) { 
+            Message Nmessage = context.bodyAsClass(Message.class); 
+            String newMessageText = Nmessage.getMessage_text(); 
+            if (!newMessageText.isBlank() && newMessageText.length() <= 255) { 
+                Message updatedMessage = mService.updateMessage(message_id, newMessageText); 
+                if (updatedMessage != null) { context.json(updatedMessage);
+                     context.status(200); } 
+                    else { context.status(400); } 
+                } 
+                else { context.status(400); } 
+            } 
+                else { context.status(400); } 
+            }
 
     private void getMessageByUserID(Context context){
         int message_id = Integer.parseInt(context.pathParam("account_id"));
